+++ minecraft/kubejs/server_scripts/Lychee.js
@@ -1,27 +1,27 @@
 ServerEvents.recipes(event => {
-    event.custom({
-        "type": "lychee:item_inside",
-        "item_in": {
-            "item": "create:refined_radiance"
-        },
-        "block_in": {
-            "blocks": ["minecraft:water"],
-        },
-        "post": [
-            {
-                "type": "place",
-                "block": "kubejs:refined_fluid"
-            },
-            {
-                "type": "drop_item",
-                "item": "create:chromatic_compound",
-                "contextual": {
-                    "type": "chance",
-                    "chance": 0.15
-                }
-            }
-        ]
-    })
+    //event.custom({
+    //    "type": "lychee:item_inside",
+    //    "item_in": {
+    //        "item": "create:refined_radiance"
+    //    },
+    //    "block_in": {
+    //        "blocks": ["minecraft:water"],
+    //    },
+    //    "post": [
+    //        {
+    //            "type": "place",
+    //            "block": "kubejs:refined_fluid"
+    //        },
+    //        {
+    //            "type": "drop_item",
+    //            "item": "create:chromatic_compound",
+    //            "contextual": {
+    //                "type": "chance",
+    //                "chance": 0.15
+    //            }
+    //        }
+    //    ]
+    //})
     event.custom({
         "type": "lychee:item_inside",
         "item_in": {
@@ -42,22 +42,22 @@ ServerEvents.recipes(event => {
             }
         ]
     })
-    event.custom({
-        "type": "lychee:block_crushing",
-        "item_in": {
-            "item": "create:chromatic_compound"
-        },
-        "post": [
-            {
-                "type": "drop_item",
-                "item": "kubejs:chromatic_shard",
-                "contextual": {
-                    "type": "chance",
-                    "chance": 1
-                }
-            }
-        ]
-    })
+    //event.custom({
+    //    "type": "lychee:block_crushing",
+    //    "item_in": {
+    //        "item": "create:chromatic_compound"
+    //    },
+    //    "post": [
+    //        {
+    //            "type": "drop_item",
+    //            "item": "kubejs:chromatic_shard",
+    //            "contextual": {
+    //                "type": "chance",
+    //                "chance": 1
+    //            }
+    //        }
+    //    ]
+    //})
     event.custom({
         "type": "lychee:item_exploding",
         "item_in": {