+++ minecraft/kubejs/server_scripts/tags.js
@@ -1,5 +1,6 @@
 ServerEvents.tags('item', event => {
     event.remove('forge:fillet_knife', 'farmersdelight:flint_knife')
+    event.remove('forge:ingots/cast_iron', 'createindustry:cast_iron_ingot')
     event.add('forge:dirtable', 'farmersdelight:rich_soil')
     //event.add('createblock:copper', ['minecraft:copper_block', 'minecraft:weathered_copper', 'minecraft:exposed_copper', 'minecraft:oxidized_copper', 'minecraft:waxed_copper_block', 'minecraft:waxed_exposed_copper', 'minecraft:waxed_weathered_copper', 'minecraft:waxed_oxidized_copper'])
     event.add('curios:back', ['create:copper_backtank', 'create:netherite_backtank', 'create_jetpack:jetpack'])
@@ -7,4 +8,6 @@ ServerEvents.tags('item', event => {
     event.add('forge:raw_materials/silver', 'kubejs:raw_silver')
     event.add('forge:ingots/silver', 'kubejs:silver_ingot')
     event.add('forge:nuggets/silver', 'kubejs:silver_nugget')
+    event.add('forge:diesel', 'createindustry:diesel')
+    event.add('forge:gasoline', 'createindustry:gasoline')
 })